requirements.txt
openai>=1.3.0
opensearch-py>=2.3.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.5.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.15.0
GitPython>=3.1.30
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
uuid
pathlib
